pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-search-results>=2.4.2",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.6",
    "langchain-mistralai>=0.2.10",
    "load-dotenv>=0.1.0",
]
